backend/requirements.txt
aiosmtplib==2.0.2 ; python_version >= "3.11" and python_version < "4.0"
annotated-types==0.5.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==3.7.1 ; python_version >= "3.11" and python_version < "4.0"
asyncpg==0.28.0 ; python_version >= "3.11" and python_version < "4.0"
blinker==1.6.2 ; python_version >= "3.11" and python_version < "4.0"
certifi==2023.7.22 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
dnspython==2.4.2 ; python_version >= "3.11" and python_version < "4.0"
ecdsa==0.18.0 ; python_version >= "3.11" and python_version < "4.0"
email-validator==2.0.0.post2 ; python_version >= "3.11" and python_version < "4.0"
fastapi-mail==1.4.1 ; python_version >= "3.11" and python_version < "4.0"
fastapi[all]==0.103.2 ; python_version >= "3.11" and python_version < "4.0"
greenlet==3.0.0 ; python_version >= "3.11" and python_version < "4.0" and (platform_machine == "win32" or platform_machine == "WIN32" or platform_machine == "AMD64" or platform_machine == "amd64" or platform_machine == "x86_64" or platform_machine == "ppc64le" or platform_machine == "aarch64")
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
httpcore==0.18.0 ; python_version >= "3.11" and python_version < "4.0"
httptools==0.6.0 ; python_version >= "3.11" and python_version < "4.0"
httpx==0.25.0 ; python_version >= "3.11" and python_version < "4.0"
idna==3.4 ; python_version >= "3.11" and python_version < "4.0"
itsdangerous==2.1.2 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==2.1.3 ; python_version >= "3.11" and python_version < "4.0"
orjson==3.9.7 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.5.0 ; python_version >= "3.11" and python_version < "4.0"
pydantic-core==2.10.1 ; python_version >= "3.11" and python_version < "4.0"
pydantic-extra-types==2.1.0 ; python_version >= "3.11" and python_version < "4.0"
pydantic-settings==2.0.3 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.4.2 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
python-jose==3.3.0 ; python_version >= "3.11" and python_version < "4.0"
python-multipart==0.0.6 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.11" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.11" and python_version < "4"
six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.11" and python_version < "4.0"
sqlalchemy==2.0.21 ; python_version >= "3.11" and python_version < "4.0"
stackprinter==0.2.10 ; python_version >= "3.11" and python_version < "4.0"
starlette==0.27.0 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.8.0 ; python_version >= "3.11" and python_version < "4.0"
ujson==5.8.0 ; python_version >= "3.11" and python_version < "4.0"
uuid7==0.1.0 ; python_version >= "3.11" and python_version < "4.0"
uvicorn[standard]==0.23.2 ; python_version >= "3.11" and python_version < "4.0"
uvloop==0.17.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.11" and python_version < "4.0"
watchfiles==0.20.0 ; python_version >= "3.11" and python_version < "4.0"
websockets==11.0.3 ; python_version >= "3.11" and python_version < "4.0"
